did it

+++ watch_local_folder.py
@@ -1,30 +1,28 @@
 import os
 import time
+from typing import Dict, Any, List
 
+from loguru import logger
 from watchdog.events import FileSystemEventHandler, FileSystemEvent
 from watchdog.observers import Observer
 
+from dotenv_reader import read_env
 from yandex_disk_api import YandexDiskAPI
+
+ignored_files: List[str] = [
+    ".DS_Store",
+]  # файлы, которые хотим игнорировать
 
 
 class MyHandler(FileSystemEventHandler):
-    """ Класс обработчика """
-    def __init__(self, cloud_api: YandexDiskAPI, sync_period: int):
+    """Класс обработчика"""
+
+    def __init__(self, cloud_api: YandexDiskAPI):
         """
         :param cloud_api: API облачного хранилища
-        :param sync_period: Период синхронизации
         """
         super().__init__()
         self.cloud_api = cloud_api
-        self.sync_period = sync_period
-
-    def set_sync_period(self, sync_period: int) -> None:
-        """
-        Установка периода синхронизации
-        :param sync_period:
-        :return:
-        """
-        self.sync_period = sync_period
 
     def on_created(self, event: FileSystemEvent) -> None:
         """
@@ -35,8 +33,10 @@ class MyHandler(FileSystemEventHandler):
         if not event.is_directory:
             local_path: str = event.src_path
             filename: str = os.path.basename(local_path)
-            self.cloud_api.load_or_reload(local_path, overwrite=False)
-            print(f'File {filename} was created')
+            if filename not in ignored_files:
+                self.cloud_api.load_or_reload(local_path, overwrite=False)
+                print(f"File {filename} was created")
+            return
 
     def on_modified(self, event: FileSystemEvent) -> None:
         """
@@ -47,8 +47,10 @@ class MyHandler(FileSystemEventHandler):
         if not event.is_directory:
             local_path: str = event.src_path
             filename: str = os.path.basename(local_path)
-            self.cloud_api.load_or_reload(local_path, overwrite=True)
-            print(f'File {filename} was modified')
+            if filename not in ignored_files:
+                self.cloud_api.load_or_reload(local_path, overwrite=True)
+                print(f"File {filename} was modified")
+            return
 
     def on_deleted(self, event: FileSystemEvent) -> None:
         """
@@ -59,24 +61,33 @@ class MyHandler(FileSystemEventHandler):
         if not event.is_directory:
             local_path: str = event.src_path
             filename: str = os.path.basename(local_path)
-            self.cloud_api.delete(os.path.basename(local_path))
-            print(f'File {filename} was deleted')
+            if filename not in ignored_files:
+                self.cloud_api.delete(os.path.basename(local_path))
+                print(f"File {filename} was deleted")
+            return
 
 
-def watch_local_folder(folder_path: str, yandex_disk_api: YandexDiskAPI, sync_period: int):
-    print('Starting up the observer')
+def watch_local_folder(local_folder: str, cloud_api: YandexDiskAPI) -> None:
+    """
+    Функция запуска мониторинга папки
+    :param local_folder: Путь к локальной папке
+    :param cloud_api: API облачного сервиса
+    :return: None
+    """
     observer: Observer = Observer()
-    observer.schedule(MyHandler(cloud_api=yandex_disk_api, sync_period=sync_period), folder_path, recursive=True)
+    observer.schedule(MyHandler(cloud_api=cloud_api), local_folder, recursive=False)
     observer.start()
+
+    env_vars: Dict[str, Any] = read_env()
+    sync_period: str = env_vars["SYNC_PERIOD"]
 
     try:
         while True:
-            time.sleep(1)
+            try:
+                time.sleep(int(sync_period))
+            except ValueError as e:
+                logger.error(f"Ошибка периода синхронизации: {e}")
+                break
     except KeyboardInterrupt:
         observer.stop()
-    observer.join()
-
-
-# if __name__ == '__main__':
-#     folder = '/Users/timofey/PycharmProjects/synchronizer'
-#     watch_local_folder(folder)
+        observer.join()
